+++ docs/source/Scheduling.rst
@@ -6,7 +6,7 @@ Scheduling
 
    Scheduler
    Condition
-   TimeScale
+   Time
    Run
    Log
