+++ docs/source/Time.rst
@@ -1,0 +1,6 @@
+Time
+====
+
+.. automodule:: psyneulink.scheduling.time
+   :members:
+   :exclude-members: random
