--- docs/source/TimeScale.rst
@@ -1,6 +1,0 @@
-TimeScale
-=========
-
-.. automodule:: psyneulink.scheduling.timescale
-   :members:
-   :exclude-members: random
