+++ psyneulink/components/mechanisms/adaptive/gating/gatingmechanism.py
@@ -171,7 +171,7 @@ from psyneulink.globals.keywords import GATING, GATING_POLICY, GATING_PROJECTION
 from psyneulink.globals.preferences.componentpreferenceset import is_pref_set
 from psyneulink.globals.preferences.preferenceset import PreferenceLevel
 from psyneulink.globals.utilities import ContentAddressableList
-from psyneulink.scheduling.timescale import TimeScale
+from psyneulink.scheduling.time import TimeScale
 
 __all__ = [
     'GatingMechanism', 'GatingMechanismError', 'GatingMechanismRegistry'
