+++ psyneulink/components/mechanisms/mechanism.py
@@ -792,7 +792,7 @@ from psyneulink.globals.log import LogLevel
 from psyneulink.globals.preferences.preferenceset import PreferenceLevel
 from psyneulink.globals.registry import register_category
 from psyneulink.globals.utilities import AutoNumber, ContentAddressableList, append_type_to_name, convert_to_np_array, iscompatible, kwCompatibilityNumeric
-from psyneulink.scheduling.timescale import TimeScale
+from psyneulink.scheduling.time import TimeScale
 
 __all__ = [
     'Mechanism_Base', 'MechanismError'
