+++ psyneulink/library/mechanisms/adaptive/learning/autoassociativelearningmechanism.py
@@ -88,7 +88,7 @@ from psyneulink.globals.keywords import AUTOASSOCIATIVE_LEARNING_MECHANISM, CONT
 from psyneulink.globals.preferences.componentpreferenceset import is_pref_set
 from psyneulink.globals.preferences.preferenceset import PreferenceLevel
 from psyneulink.globals.utilities import is_numeric, parameter_spec
-from psyneulink.scheduling.timescale import TimeScale
+from psyneulink.scheduling.time import TimeScale
 
 __all__ = [
     'AutoAssociativeLearningMechanism', 'AutoAssociativeLearningMechanismError', 'DefaultTrainingMechanism',
