+++ psyneulink/scheduling/__init__.py
@@ -6,12 +6,12 @@ https://princetonuniversity.github.io/PsyNeuLink/Scheduling.html
 
 from . import condition
 from . import scheduler
-from . import timescale
+from . import time
 
 from .condition import *
 from .scheduler import *
-from .timescale import *
+from .time import *
 
 __all__ = list(condition.__all__)
 __all__.extend(scheduler.__all__)
-__all__.extend(timescale.__all__)
+__all__.extend(time.__all__)
