+++ tests/scheduling/test_scheduler.py
@@ -10,7 +10,7 @@ from psyneulink.scheduling.condition import AfterNCalls, AfterNTrials, AfterPass
     BeforePass, \
     EveryNCalls, EveryNPasses, JustRan, WhenFinished
 from psyneulink.scheduling.scheduler import Scheduler
-from psyneulink.scheduling.timescale import TimeScale
+from psyneulink.scheduling.time import TimeScale
 
 logger = logging.getLogger(__name__)
